+++ XcodeML-Exc-Tools/src/exc/xmpF/XMPenv.java
@@ -294,12 +294,13 @@ out:{
       for(XobjectDef def = current_def.getDef(); def != null;
 	  def = def.getParent()){
         Xobject id_list = def.getDef().getArg(1);
-	if (id_list == null) continue;
-        for(Xobject i: (XobjList)id_list){
-	  if(i.getName().equals(name)){
-	    id = (Ident)i;
-            break out;
-	  }
+        if (id_list != null) {
+          for(Xobject i: (XobjList)id_list){
+            if(i.getName().equals(name)){
+              id = (Ident)i;
+              break out;
+            }
+          }
         }
       }
       return null;