+++ XcodeML-Exc-Tools/src/exc/xmpF/XMPtransCoarrayRun.java
@@ -485,6 +485,8 @@ public class XMPtransCoarrayRun
     localCoarrays = new ArrayList<XMPcoarray>();
     moduleCoarrays = new ArrayList<XMPcoarray>();
     Xobject idList = def.getFuncIdList();
+    if (idList == null)
+      idList = new XobjList(Xcode.ID_LIST);
 
     /* divide coarrays into localCoarrays and moduleCoarrays
      */
@@ -807,7 +809,7 @@ public class XMPtransCoarrayRun
     --------------------------------------------
       subroutine EX1
         real :: V1(1:10,1:20)                                        ! f. f1.
-        common /xmpf_COARRAY_EX1/ V1                                 ! f4.
+        common /xmpf_COARRAY_EX1/ V1                                 ! c4.
         integer(8) :: descptr_V1                                     ! a.
         common /xmpf_descptr_EX1/ descptr_V1                         ! a1.
         ...
@@ -926,7 +928,7 @@ public class XMPtransCoarrayRun
       // f1. remove SAVE attributes from declarations of coarrays
       removeSaveAttr(staticLocalCoarrays);
       if (version == 4) {
-        // f4. generate common block for data
+        // c4. generate common block for data
         genCommonBlockForStaticCoarrays(staticLocalCoarrays);
       }
     }
@@ -1332,6 +1334,16 @@ public class XMPtransCoarrayRun
 
         if (allocated(V3)) write(*,*) "yes"       ! keep 'allocated' ! l2.
     --------------------------------------------
+
+    common and special handling for the COARRAY directive
+    (for the coarray pass that executes after the XMP pass)
+    --------------------------------------------
+        call xmpf_coarray_set_nodes ( a , XMP_DESC_p )
+    --------------------------------------------
+    output (commonly)
+    --------------------------------------------
+        call xmpf_coarray_set_nodes ( xmpf_descptr_a , XMP_DESC_p )  ! r1.
+    --------------------------------------------
   */
   private void transExecPart() {
 
@@ -1356,6 +1368,9 @@ public class XMPtransCoarrayRun
 
     // l2. fake intrinsic 'allocatable' (allocatable coarrays only)
     replaceIntrinsicCalls2(visibleCoarrays);
+
+    // r1. replace argument of runtime call
+    replaceArgumentOfRuntimeCall(visibleCoarrays);
 
     // i. initialization/finalization for auto-syncall and auto-deallocate
     //    and initialization of descPtr (only Ver.6)
@@ -2692,6 +2707,13 @@ public class XMPtransCoarrayRun
     }
 
     Xobject tag;
+    /////////////////////////////////
+    ////  for issue #586
+    ////   The trigger of the auto-deallocation should be not an allocate statement
+    ////   but the declaration of the allocatable coarray.
+    //System.out.println("@@@@ "+(coarray.wasMovedFromModule())+(coarray.def != def)+
+    //             (coarray.usesMalloc()));
+    /////////////////////////////////
     if (coarray.wasMovedFromModule() || coarray.def != def ||
         !coarray.usesMalloc()) {
       // coarray is originally defined in a use-associated module or
@@ -3001,6 +3023,42 @@ public class XMPtransCoarrayRun
         _replaceAllocatedWithAssociated(xobj, coarrays);
     }
   }
+
+
+  //-----------------------------------------------------
+  //  TRANSLATION r1.
+  //  - replace argument of runtime call for COARRAY directive
+  //-----------------------------------------------------
+  //
+  private void replaceArgumentOfRuntimeCall(ArrayList<XMPcoarray> coarrays) {
+    XobjectIterator xi = new topdownXobjectIterator(def.getFuncBody());
+    for (xi.init(); !xi.end(); xi.next()) {
+      Xobject xobj = xi.getXobject();
+      if (xobj == null)
+        continue;
+      if (xobj.Opcode() != Xcode.FUNCTION_CALL ||
+          xobj.getArg(0).Opcode() == Xcode.MEMBER_REF)
+        continue;
+
+      /*  Replace coarray argument with the descriptor
+       *  if the function name is xmpf_coarray_set_nodes.
+       */
+      String fname = xobj.getArg(0).getString();
+      if (fname.equals(XMPcoarray.SET_NODES_NAME)) {
+        for (XobjArgs args = ((XobjList)xobj.getArg(1)).getArgs();
+             args != null; args = args.nextArgs()) {
+          Xobject arg = args.getArg();
+          XMPcoarray coarray = _findCoarrayInCoarrays(arg, coarrays);
+          if (coarray != null) {
+            // found the argument to be replaced
+            Ident descPtr = coarray.getDescPointerId();
+            args.setArg(descPtr);
+          }
+        }
+      }
+    }
+  }
+
 
 
   // not used currently.