+++ tests/XMP/others/F/python/SPMD/test.f90
@@ -1,0 +1,7 @@
+subroutine hello(a, b) bind(C)
+  implicit none
+  include 'xmp_lib.h'
+  integer(8), intent(in) :: a(3), b(3)
+
+  write(*,*) xmp_node_num(), a(:), b(:)
+end subroutine hello_2